# Improve speed  of getting local OCAT data