@taldcroft 2022-05-14
I fixed the thing about using an observer name, though that comments seems to be gone now.